use float instead of double

+++ aabb.h
@@ -53,7 +53,7 @@ class aabb {
 
         for (int axis = 0; axis < 3; axis++) {
             const interval& ax = axis_interval(axis);
-            const double adinv = 1.0 / ray_dir[axis];
+            const float adinv = 1.0 / ray_dir[axis];
 
             auto t0 = (ax.min - ray_orig[axis]) * adinv;
             auto t1 = (ax.max - ray_orig[axis]) * adinv;
@@ -88,7 +88,7 @@ class aabb {
     void pad_to_minimums() {
         // Adjust the AABB so that no side is narrower than some delta, padding if necessary.
 
-        double delta = 0.0001;
+        float delta = 0.0001;
         if (x.size() < delta) x = x.expand(delta);
         if (y.size() < delta) y = y.expand(delta);
         if (z.size() < delta) z = z.expand(delta);
